Feature: add tree merge to laws_parser
