+++ rag/app/laws.py
@@ -23,74 +23,11 @@ from docx import Document
 from api.db import ParserType
 from deepdoc.parser.utils import get_text
 from rag.nlp import bullets_category, remove_contents_table, hierarchical_merge, \
-    make_colon_as_title, tokenize_chunks, docx_question_level
-from rag.nlp import rag_tokenizer
+    make_colon_as_title, tokenize_chunks, docx_question_level, tree_merge
+from rag.nlp import rag_tokenizer, Node
 from deepdoc.parser import PdfParser, DocxParser, PlainParser, HtmlParser
 
-class Node:
-    def __init__(self, level, depth=-1, texts=None):
-        self.level = level
-        self.depth = depth
-        self.texts = texts if texts is not None else []  # 存放内容
-        self.children = []  # 子节点
 
-    def add_child(self, child_node):
-        self.children.append(child_node)
-
-    def get_children(self):
-        return self.children
-
-    def get_level(self):
-        return self.level
-
-    def get_texts(self):
-        return self.texts
-
-    def set_texts(self, texts):
-        self.texts = texts
-
-    def add_text(self, text):
-        self.texts.append(text)
-
-    def clear_text(self):
-        self.texts = []
-
-    def __repr__(self):
-        return f"Node(level={self.level}, texts={self.texts}, children={len(self.children)})"
-
-    def build_tree(self, lines):
-        stack = [self]  
-        for line in lines:
-            level, text = line
-            node = Node(level=level, texts=[text])
-
-            if level <= self.depth or self.depth == -1:
-                while stack and level <= stack[-1].get_level():
-                    stack.pop()
-
-                stack[-1].add_child(node)
-                stack.append(node)
-            else:
-                stack[-1].add_text(text)
-        return self  
-
-    def get_tree(self):
-        tree_list = []  
-        self._dfs(self, tree_list, 0, [])
-        return tree_list
-
-    def _dfs(self, node, tree_list, current_depth, titles):
-
-        if node.get_texts():
-            if 0 < node.get_level() < self.depth:
-                titles.extend(node.get_texts())
-            else:
-                combined_text = ["\n".join(titles + node.get_texts())]
-                tree_list.append(combined_text)
-
-
-        for child in node.get_children():
-            self._dfs(child, tree_list, current_depth + 1, titles.copy())
 
 
 
@@ -121,35 +58,37 @@ class Docx(DocxParser):
         return [line for line in lines if line]
 
     def __call__(self, filename, binary=None, from_page=0, to_page=100000):
-        self.doc = Document(
-            filename) if not binary else Document(BytesIO(binary))
-        pn = 0
-        lines = []
-        level_set = set()
-        bull = bullets_category([p.text for p in self.doc.paragraphs])
-        for p in self.doc.paragraphs:
-            if pn > to_page:
-                break
-            question_level, p_text = docx_question_level(p, bull)
-            if not p_text.strip("\n"):
-                continue
-            lines.append((question_level, p_text))
-            level_set.add(question_level)
-            for run in p.runs:
-                if 'lastRenderedPageBreak' in run._element.xml:
-                    pn += 1
+            self.doc = Document(
+                filename) if not binary else Document(BytesIO(binary))
+            pn = 0
+            lines = []
+            level_set = set()
+            bull = bullets_category([p.text for p in self.doc.paragraphs])
+            for p in self.doc.paragraphs:
+                if pn > to_page:
+                    break
+                question_level, p_text = docx_question_level(p, bull)
+                if not p_text.strip("\n"):
                     continue
-                if 'w:br' in run._element.xml and 'type="page"' in run._element.xml:
-                    pn += 1
+                lines.append((question_level, p_text))
+                level_set.add(question_level)
+                for run in p.runs:
+                    if 'lastRenderedPageBreak' in run._element.xml:
+                        pn += 1
+                        continue
+                    if 'w:br' in run._element.xml and 'type="page"' in run._element.xml:
+                        pn += 1
 
-        sorted_levels = sorted(level_set)
+            sorted_levels = sorted(level_set)
 
-        h2_level = sorted_levels[1] if len(sorted_levels) > 1 else 1    
+            h2_level = sorted_levels[1] if len(sorted_levels) > 1 else 1
+            h2_level = sorted_levels[-2] if h2_level == sorted_levels[-1] and len(sorted_levels) > 2 else h2_level
+            
+            root = Node(level=0, depth=h2_level, texts=[])
+            root.build_tree(lines)
 
-        root = Node(level=0, depth=h2_level, texts=[])
-        root.build_tree(lines)
+            return [("\n").join(element) for element in root.get_tree() if element]
 
-        return [("\n").join(element) for element in root.get_tree() if element]
 
     def __str__(self) -> str:
         return f'''
@@ -215,7 +154,7 @@ def chunk(filename, binary=None, from_page=0, to_page=100000,
         chunks = Docx()(filename, binary)
         callback(0.7, "Finish parsing.")
         return tokenize_chunks(chunks, doc, eng, None)
-
+    
     elif re.search(r"\.pdf$", filename, re.IGNORECASE):
         pdf_parser = Pdf()
         if parser_config.get("layout_recognize", "DeepDOC") == "Plain Text":
@@ -224,7 +163,7 @@ def chunk(filename, binary=None, from_page=0, to_page=100000,
                                     from_page=from_page, to_page=to_page, callback=callback)[0]:
             sections.append(txt + poss)
 
-    elif re.search(r"\.txt$", filename, re.IGNORECASE):
+    elif re.search(r"\.(txt|md|markdown|mdx)$", filename, re.IGNORECASE):
         callback(0.1, "Start to parse.")
         txt = get_text(filename, binary)
         sections = txt.split("\n")
@@ -255,13 +194,16 @@ def chunk(filename, binary=None, from_page=0, to_page=100000,
 
     make_colon_as_title(sections)
     bull = bullets_category(sections)
-    chunks = hierarchical_merge(bull, sections, 5)
-    if not chunks:
+    res = tree_merge(bull, sections, 2)
+
+
+    if not res:
         callback(0.99, "No chunk parsed out.")
 
-    return tokenize_chunks(["\n".join(ck)
-                           for ck in chunks], doc, eng, pdf_parser)
+    return tokenize_chunks(res, doc, eng, pdf_parser)
 
+    # chunks = hierarchical_merge(bull, sections, 5)
+    #     return tokenize_chunks(["\n".join(ck)for ck in chunks], doc, eng, pdf_parser)
 
 if __name__ == "__main__":
     import sys
